implementation of magic bitboards

+++ include/evaluation.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "board.h"
 
-double evaluate(const Board& board);
+int evaluate(const Board& board);
